+++ src/type-functions.php
@@ -53,6 +53,7 @@ use GanbaroDigital\MissingBits\TypeInspectors\GetStringDuckTypes;
 use GanbaroDigital\MissingBits\TypeInspectors\GetStringTypes;
 use GanbaroDigital\MissingBits\TypeInspectors\IsList;
 use GanbaroDigital\MissingBits\TypeInspectors\IsListyObject;
+use GanbaroDigital\MissingBits\TypeInspectors\IsStringy;
 use GanbaroDigital\MissingBits\TypeInspectors\StripNamespace;
 
 /**
@@ -243,25 +244,7 @@ function is_listy_object($list)
  */
 function is_stringy($item)
 {
-    // PHP will auto-convert these to strings without generating
-    // any errors
-    if (is_string($item) || is_int($item) || is_double($item)) {
-        return true;
-    }
-
-    // depends if the object has the __toString() method or not
-    if (is_object($item)) {
-        return method_exists($item, '__toString');
-    }
-
-    // there's no point turning these into strings
-    //
-    // NULL
-    // array
-    // boolean
-    // callable
-    // resource
-    return false;
+    return IsStringy::check($item);
 }
 
 /**